Initial implementation
This is a first attempt for implementing the mars rover. I like the
simplicity of this approach. The MarsRover class is not that big and at
this stage it's difficult to make assumptions about the right
abstractions. This is a good enough implementation for the moment.

What I don't like is that the MarsRover class has too many
responsibilities and it should be split. Probably a GridBuilder and
Position classes can be extracted. A main entry point also can be
created and then the MarsRover could have only logic about the rover and
how he moves.