Extract utility classes
Extracts some classes that should be separate because they are helpers
to build the world and are not connected with the rover.

Extracts the driving module of the rover to a separate class.

Fixes some difficult to catch bugs and allows to build grids with
different sizes and automatically to pick the changes in the MarsRover
class.

Adds tests for the new classes because now it's easier to test the
functionality of the classes in separation.